kill JSONCell in favor of a custom display method and style overrides at TableNG
