+++ packages/grafana-schema/src/common/table.cue
@@ -55,7 +55,6 @@ TableImageCellOptions: {
 // Show data links in the cell
 TableDataLinksCellOptions: {
 	type: TableCellDisplayMode & "data-links"
-  // wrapText?: bool
 } @cuetsy(kind="interface")
 
 // Show actions in the cell
