+++ packages/grafana-ui/src/components/Table/TableNG/Cells/renderers.tsx
@@ -11,7 +11,6 @@ import { BarGaugeCell } from './BarGaugeCell';
 import { DataLinksCell } from './DataLinksCell';
 import { GeoCell } from './GeoCell';
 import { ImageCell } from './ImageCell';
-import { JSONCell } from './JSONCell';
 import { PillCell } from './PillCell';
 import { SparklineCell } from './SparklineCell';
 
@@ -42,10 +41,6 @@ const SPARKLINE_RENDERER: TableCellRenderer = (props) => (
     theme={props.theme}
     width={props.width}
   />
-);
-
-const JSON_RENDERER: TableCellRenderer = (props) => (
-  <JSONCell value={props.value} field={props.field} rowIdx={props.rowIdx} />
 );
 
 const GEO_RENDERER: TableCellRenderer = (props) => (
@@ -86,7 +81,7 @@ const CUSTOM_RENDERER: TableCellRenderer = (props) => {
 const CELL_RENDERERS: Record<TableCellOptions['type'], TableCellRenderer> = {
   [TableCellDisplayMode.Sparkline]: SPARKLINE_RENDERER,
   [TableCellDisplayMode.Gauge]: GAUGE_RENDERER,
-  [TableCellDisplayMode.JSONView]: JSON_RENDERER,
+  [TableCellDisplayMode.JSONView]: AUTO_RENDERER,
   [TableCellDisplayMode.Image]: IMAGE_RENDERER,
   [TableCellDisplayMode.DataLinks]: DATA_LINKS_RENDERER,
   [TableCellDisplayMode.Actions]: ACTIONS_RENDERER,
@@ -121,12 +116,7 @@ export function getAutoRendererResult(field: Field): TableCellRenderer {
     const firstValue = field.values[0];
     if (isDataFrame(firstValue) && isTimeSeriesFrame(firstValue)) {
       return SPARKLINE_RENDERER;
-    } else {
-      return JSON_RENDERER;
     }
-  }
-  if (field.type === FieldType.other) {
-    return JSON_RENDERER;
   }
   return AUTO_RENDERER;
 }
